UPDATE [sc-1162] - moved utils into lib
- also updated components to use latest versions of @owl packages
