+++ packages/button/src/Default/Default.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Button, ThemeProvider } from 'react-bootstrap';
 import { ButtonDefaultProps } from './Default.types';
 import * as styleMod from './styles.module.scss';
-import { createLocalProps, themePrefixesProps } from '@owlui/utils';
+import { createLocalProps, themePrefixesProps } from '@owlui/lib/src/utils';
 
 export const Component = (props: ButtonDefaultProps) => {
   const themePrefixes: themePrefixesProps = {};
