+++ packages/icons/package.json
@@ -31,6 +31,6 @@
   "dependencies": {
     "@owlui/theme": "^0.0.24",
     "@owlui/typings": "^0.0.24",
-    "@owlui/utils": "^0.0.24"
+    "@owlui/lib": "^0.0.24"
   }
 }
