+++ packages/switchToggle/src/Default/Default.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Form, ThemeProvider } from 'react-bootstrap';
 import { SwitchToggleDefaultProps } from './Default.types';
 import * as styleMod from './styles.module.scss';
-import { createLocalProps } from '@owlui/utils';
+import { createLocalProps } from '@owlui/lib/src/utils';
 
 export const Component = (props: SwitchToggleDefaultProps) => {
   const baseClass = 'switchToggle';
