+++ packages/table/src/Default/Default.tsx
@@ -6,7 +6,7 @@ import {
   TableRowItem,
 } from './Default.types';
 import * as styleMod from './styles.module.scss';
-import { createLocalProps } from '@owlui/utils';
+import { createLocalProps } from '@owlui/lib/src/utils';
 
 export const Component = (props: TableDefaultProps) => {
   const baseClass = 'table';
