+++ scripts/components/templates/lib/index.hbs
@@ -1,4 +1,4 @@
-export * from '../../utils/src/index';
+export * from './utils';
 {{#with . as | data |}}
 {{#each data.components as | comp |}}
 export * from '{{comp.path}}';
