+++ scripts/components/templates/option/component-default.hbs
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { ThemeProvider } from 'react-bootstrap';
 import { {{{cap}}}{{{optionPas}}}Props } from './{{{optionPas}}}.types';
 import * as styleMod from './styles.module.scss';
-import { createLocalProps, themePrefixesProps } from '@owlui/utils';
+import { createLocalProps, themePrefixesProps } from '@owlui/lib/src/utils';
 
 export const Component = (props: {{{cap}}}{{{optionPas}}}Props) => {
   const themePrefixes: themePrefixesProps = {};
