--- utils/.npmignore
@@ -1,2 +1,0 @@
-*.ts
-!*.d.ts
