--- utils/CHANGELOG.md
@@ -1,64 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [0.0.24](https://github.com/EEBOS/owl-ui/compare/v0.0.23...v0.0.24) (2022-06-30)
-
-**Note:** Version bump only for package @owlui/utils
-
-
-
-
-
-## [0.0.23](https://github.com/EEBOS/owl-ui/compare/v0.0.22...v0.0.23) (2022-06-29)
-
-**Note:** Version bump only for package @owlui/utils
-
-
-
-
-
-## [0.0.22](https://github.com/EEBOS/owl-ui/compare/v0.0.21...v0.0.22) (2022-06-06)
-
-**Note:** Version bump only for package @owlui/utils
-
-
-
-
-
-## [0.0.20](https://github.com/EEBOS/owl-ui/compare/v0.0.19...v0.0.20) (2022-05-13)
-
-**Note:** Version bump only for package @owlui/utils
-
-
-
-
-
-## [0.0.19](https://github.com/EEBOS/owl-ui/compare/v0.0.18...v0.0.19) (2022-04-11)
-
-**Note:** Version bump only for package @owlui/utils
-
-
-
-
-
-## [0.0.18](https://github.com/EEBOS/owl-ui/compare/v0.0.17...v0.0.18) (2022-04-05)
-
-**Note:** Version bump only for package @owlui/utils
-
-
-
-
-
-## [0.0.17](https://github.com/EEBOS/owl-ui/compare/v0.0.16...v0.0.17) (2022-04-05)
-
-**Note:** Version bump only for package @owlui/utils
-
-
-
-
-
-## [0.0.16](https://github.com/EEBOS/owl-ui/compare/v0.0.15...v0.0.16) (2022-04-01)
-
-**Note:** Version bump only for package @owlui/utils
