--- utils/README.md
@@ -1,11 +1,0 @@
-# `utils`
-
-> TODO: description
-
-## Usage
-
-```
-const utils = require('utils');
-
-// TODO: DEMONSTRATE API
-```
