--- utils/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@owlui/utils",
-  "version": "0.0.24",
-  "description": "",
-  "author": "",
-  "homepage": "https://github.com/EEBOS/owl-ui/tree/main/packages/utils#readme",
-  "license": "MIT",
-  "source": "./src/index.ts",
-  "main": "./dist/owl.utils.js",
-  "types": "./dist/owl.utils.d.ts",
-  "directories": {
-    "lib": "lib",
-    "test": "__tests__"
-  },
-  "files": [
-    "dist",
-    "src"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/EEBOS/owl-ui.git",
-    "directory": "packages/utils"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "isLibrary": true,
-  "outputFormat": "commonjs",
-  "peerDependencies": {
-    "react": "^18.0.0"
-  },
-  "dependencies": {
-    "@owlui/theme": "^0.0.24",
-    "@owlui/typings": "^0.0.20"
-  }
-}
