--- utils/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../tsconfig.base.json",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "../src/*": ["./*"]
-    },
-    "types": ["custom-globals", "static-assets"],
-    "typeRoots": ["../node_modules/@types", "../typings"]
-  },
-  "exclude": ["dist/*"]
-}
